Added support for both bash and jenkins for terraform tfvars

+++ scripts/0-bootstrap/bootstrap.sh
@@ -18,8 +18,12 @@ echo Removing any past deployment terraform.tfvars files
 PAST_TF_VARS_FILE=./terraform.tfvars
 [ -f $PAST_TF_VARS_FILE ] && { echo "Removing past deployment file "; rm $PAST_TF_VARS_FILE; } || echo "No past deployment files found"
 
-echo Copying in variable file
+echo Copying in variable file for Jenkins deployment
 TF_VARS_FILE=../../../scripts/0-bootstrap/terraform.auto.tfvars.json
+[ -f $TF_VARS_FILE ] && { echo "Copying terraform.tfvars"; cp $TF_VARS_FILE .; } || { echo "can't find $TF_VARS_FILE"; exit 1; }
+
+echo Copying in variable file for bash deployment
+TF_VARS_FILE=../../../scripts/0-bootstrap/terraform.tfvars
 [ -f $TF_VARS_FILE ] && { echo "Copying terraform.tfvars"; cp $TF_VARS_FILE .; } || { echo "can't find $TF_VARS_FILE"; exit 1; }
 
 echo Running terraform init/plan/apply
